Second structure.
