--- server/src/middleware/validate.ts
@@ -1,22 +1,0 @@
-import { Request, Response, NextFunction } from 'express';
-import { ValidationChain, validationResult } from 'express-validator';
-
-export const validateRequest = (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): void => {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    res.status(400).json({ errors: errors.array() });
-    return;
-  }
-  next();
-};
-
-export const validate = (validations: ValidationChain[]) => {
-  return async (req: Request, res: Response, next: NextFunction) => {
-    await Promise.all(validations.map(validation => validation.run(req)));
-    validateRequest(req, res, next);
-  };
-};
