+++ server/src/shared/middleware/errorHandler.ts
