+++ server/src/shared/utils/errors.ts
